+++ src/gui/shell.h
@@ -42,7 +42,7 @@ signals:
 public slots:
 	void handleNeovimNotification(const QByteArray &name, const QVariantList& args);
 	void resizeNeovim(const QSize&);
-	void resizeNeovim(uint64_t n_cols, uint64_t n_rows);
+	void resizeNeovim(int n_cols, int n_rows);
 	bool setGuiFont(const QString& fdesc);
 
 protected slots:
